fix(inspection): Codex review 修复 5 个必改项
1. runNow() 传入 pool + logger（之前无参数会 TypeError）
2. SELECT 列名对齐 DDL（metric_key/detail → type/title/description/...）
3. days_since 转 ::int（PostgreSQL interval 不能直接比较整数）
4. persistResult 加事务（BEGIN/COMMIT/ROLLBACK 防半写入）
5. 前端 inspected_at → created_at（对齐 API 返回字段）

+++ apps/gateway/services/inspection/engine.js
@@ -171,7 +171,7 @@ async function detectNewProductUnderperform(pool, anomalies) {
     sku_total AS (
       SELECT d.sku,
              f.first_date,
-             (current_date - f.first_date) AS days_since,
+             (current_date - f.first_date)::int AS days_since,
              sum(${salesSumExpr}) AS total_qty
       FROM ${SALES_DAILY_TABLE} d
       JOIN sku_first_seen f ON f.sku = d.sku
